# swagger/admin/v1/schema.yaml
...
openapi: 3.0.3
info:
  title: FleetYards.net Command API
  version: v1
  license:
    name: GNU General Public License v3.0
    url: https://github.com/fleetyards/fleetyards/blob/main/LICENSE
  x-logo:
    url: https://fleetyards.net/docs/logo.png
    altText: FleetYards.net logo
servers:
- url: http://admin.fleetyards.test/api/v1
  description: Dev Server
- url: http://admin.fleetyards.test/api/v1
  description: Production Server
- url: https://admin.fleetyards.dev/api/v1
  description: Staging Server
security:
- SessionCookie: []
paths:
...
  "/components/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      description: Component id
      required: true
    get:
      summary: Component Detail
      operationId: component
      tags:
      - Components
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Component"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '403':
          description: forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/manufacturers/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      description: Manufacturer id
      required: true
    get:
      summary: Manufacturer Detail
      operationId: manufacturer
      tags:
      - Manufacturers
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Manufacturer"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '403':
          description: forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
components:
...
  schemas:
...
    CargoHold:
      type: object
      properties:
        dimensions:
          "$ref": "#/components/schemas/CargoHoldDimension"
        capacity:
          type: integer
        maxContainerSize:
          "$ref": "#/components/schemas/CargoHoldContainerSize"
        limits:
          type: object
          properties:
            min:
              "$ref": "#/components/schemas/CargoHoldLimit"
            max:
              "$ref": "#/components/schemas/CargoHoldLimit"
          additionalProperties: false
          required:
          - min
      additionalProperties: false
      required:
      - dimensions
      - capacity
      - maxContainerSize
      - limits
      title: CargoHold
    CargoHoldContainerSize:
      type: object
      properties:
        size:
          type: integer
        dimensions:
          "$ref": "#/components/schemas/CargoHoldDimension"
      additionalProperties: false
      required:
      - size
      - dimensions
      title: CargoHoldContainerSize
    CargoHoldDimension:
      type: object
      properties:
        x:
          type: number
        "y":
          type: number
        z:
          type: number
      additionalProperties: false
      required:
      - x
      - "y"
      - z
      title: CargoHoldDimension
    CargoHoldLimit:
      type: object
      properties:
        dimensions:
          "$ref": "#/components/schemas/CargoHoldDimension"
        capacity:
          type: integer
      additionalProperties: false
      required:
      - dimensions
      - capacity
      title: CargoHoldLimit
...
    Component:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        scKey:
          type: string
        scRef:
          type: string
        hidden:
          type: boolean
        category:
          type: string
        type:
          "$ref": "#/components/schemas/ComponentTypeEnum"
        subType:
          type: string
        inventoryConsumption:
          type: string
        grade:
          type: string
        gradeLabel:
          type: string
        size:
          type: string
        class:
          "$ref": "#/components/schemas/ComponentClassEnum"
        itemClass:
          "$ref": "#/components/schemas/ComponentItemClassEnum"
        itemClassLabel:
          type: string
        availability:
          type: object
          properties:
            boughtAt:
              type: array
              items:
                "$ref": "#/components/schemas/ItemPrice"
            soldAt:
              type: array
              items:
                "$ref": "#/components/schemas/ItemPrice"
          additionalProperties: false
          required:
          - boughtAt
          - soldAt
        manufacturer:
          "$ref": "#/components/schemas/Manufacturer"
        media:
          type: object
          properties:
            storeImage:
              "$ref": "#/components/schemas/MediaFile"
          additionalProperties: false
        typeData:
          anyOf:
          - "$ref": "#/components/schemas/ComponentQuantumDrive"
          - "$ref": "#/components/schemas/CargoHold"
          - "$ref": "#/components/schemas/FuelTank"
          - "$ref": "#/components/schemas/ComponentThruster"
        hardpoints:
          type: array
          items:
            "$ref": "#/components/schemas/Hardpoint"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - slug
      - hidden
      - availability
      - media
      - createdAt
      - updatedAt
      title: Component
    ComponentQuantumDrive:
      type: object
      properties:
        fuelRate:
          type: number
        jumpRange:
          type: number
        standardJump:
          "$ref": "#/components/schemas/ComponentQuantumDriveJump"
        splineJump:
          "$ref": "#/components/schemas/ComponentQuantumDriveJump"
      additionalProperties: false
      required:
      - fuelRate
      - jumpRange
      - standardJump
      - splineJump
      title: ComponentQuantumDrive
    ComponentQuantumDriveJump:
      type: object
      properties:
        speed:
          type: number
        stage1AccelerationRate:
          type: number
        stage2AccelerationRate:
          type: number
        spoolUpTime:
          type: number
        cooldown:
          type: number
      additionalProperties: false
      required:
      - speed
      - stage1AccelerationRate
      - stage2AccelerationRate
      - spoolUpTime
      - cooldown
      title: ComponentQuantumDriveJump
    ComponentThruster:
      type: object
      properties:
        thrustCapacity:
          type: number
        thrusterType:
          type: string
        thrusterClass:
          "$ref": "#/components/schemas/ThrusterClassEnum"
        fuelBurnRatePer10KNewton:
          type: number
      additionalProperties: false
      required:
      - capacity
      - thrusterType
      - fuelBurnRatePer10KNewton
      title: ComponentThruster
...
    ComponentClassEnum:
      type: string
      enum:
      - RSIModular
      - RSIWeapon
      - RSIAvionic
      - RSIPropulsion
      - RSIThruster
      title: ComponentClassEnum
    ComponentItemClassEnum:
      type: string
      enum:
      - stealth
      - civilian
      - industrial
      - military
      - competition
      title: ComponentItemClassEnum
    ComponentTypeEnum:
      type: string
      enum:
      - Seat
      - Door
      - MainThruster
      - Armor
      - WeaponDefensive
      - ManneuverThruster
      - CargoGrid
      - Cargo
      - EMP
      - Paints
      - TurretBase
      - ToolArm
      - Turret
      - Misc
      - Module
      - ControlPanel
      - WeaponGun
      - Container
      - TractorBeam
      - UtilityTurret
      - TowingBeam
      - MissileController
      - Lightgroup
      - Battery
      - CapacitorAssignmentController
      - CommsController
      - CoolerController
      - DoorController
      - EnergyController
      - FlightController
      - MiningController
      - BombLauncher
      - Bomb
      - FuelController
      - SalvageController
      - ShieldController
      - TargetSelector
      - WeaponController
      - WheeledController
      - Cooler
      - DockingAnimator
      - Button
      - Sensor
      - Usable
      - ExternalFuelTank
      - FuelTank
      - WeaponMining
      - SalvageHead
      - JumpDrive
      - FuelIntake
      - MiningModifier
      - Missile
      - AttachedPart
      - LifeSupportGenerator
      - SpaceMine
      - MissileLauncher
      - DockingCollar
      - PowerPlant
      - QuantumFuelTank
      - QuantumInterdictionGenerator
      - QuantumDrive
      - Relay
      - Radar
      - Scanner
      - SalvageFillerStation
      - Shield
      - SeatDashboard
      - Display
      - SelfDestruct
      - WeaponAttachment
      - SalvageFieldEmitter
      - SalvageFieldSupporter
      - SalvageInternalStorage
      title: ComponentTypeEnum
...
    HardpointCategoryEnum:
      type: string
      enum:
      - radar
      - computers
      - scanners
      - powerplant
      - cooler
      - shieldgenerator
      - module
      - salvagefillerstation
      - fueltanks
      - fuel_intakes
      - quantumdrive
      - jumpdrive
      - main_thrusters
      - retro_thrusters
      - vtol_thrusters
      - maneuvering_thrusters
      - weapons
      - weapon_mounts
      - turret
      - missile_racks
      - bombcompartments
      - quantumenforcementdevice
      - emp
      - salvagemunching
      - armor
      - countermeasures
      - selfdestruct
      - lifesupport
      - batteries
      - utility
      - seat
      - relay
      - paints
      - doors
      - cargogrid
      - inventory
      - controller
      - unknown
      title: HardpointCategoryEnum
    HardpointGroupEnum:
      type: string
      enum:
      - avionic
      - system
      - propulsion
      - thruster
      - weapon
      - defense
      - auxiliary
      - seat
      - relay
      - other
      - unknown
      title: HardpointGroupEnum
    HardpointSourceEnum:
      type: string
      enum:
      - ship_matrix
      - game_files
      title: HardpointSourceEnum
...
    ItemPriceItemTypeEnum:
      type: string
      enum:
      - Model
      - ModelModule
      - ModelPaint
      - Component
      title: ItemPriceItemTypeEnum
    ItemPriceTimeRangeEnum:
      type: string
      enum:
      - 1-day
      - 3-days
      - 7-days
      - 30-days
      title: ItemPriceTimeRangeEnum
    ItemPriceTypeEnum:
      type: string
      enum:
      - buy
      - sell
      - rental
      title: ItemPriceTypeEnum
...
    ThrusterClassEnum:
      type: string
      enum:
      - main
      - retro
      - vtol
      - mav
      title: ThrusterClassEnum
...
    FuelTank:
      type: object
      properties:
        capacity:
          type: number
      additionalProperties: false
      required:
      - capacity
      title: FuelTank
...
    Hardpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        group:
          "$ref": "#/components/schemas/HardpointGroupEnum"
        groupKey:
          type: string
        matrixKey:
          type: string
        category:
          "$ref": "#/components/schemas/HardpointCategoryEnum"
        name:
          type: string
        minSize:
          type: integer
        maxSize:
          type: integer
        source:
          "$ref": "#/components/schemas/HardpointSourceEnum"
        types:
          type: array
          items:
            type: string
        details:
          type: string
        component:
          "$ref": "#/components/schemas/Component"
        hardpoints:
          type: array
          items:
            "$ref": "#/components/schemas/Hardpoint"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - createdAt
      - updatedAt
      title: Hardpoint
...
    MediaFile:
      type: object
      properties:
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
        url:
          type: string
          format: uri
        smallUrl:
          type: string
          format: uri
        mediumUrl:
          type: string
          format: uri
        largeUrl:
          type: string
          format: uri
        xlargeUrl:
          type: string
          format: uri
        width:
          type: integer
        height:
          type: integer
        uploadedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - name
      - contentType
      - size
      - url
      - uploadedAt
      title: MediaFile
...
    StandardError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
      required:
      - code
      - message
      title: StandardError
...
    ItemPrice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        price:
          type: number
        timeRange:
          "$ref": "#/components/schemas/ItemPriceTimeRangeEnum"
        priceType:
          "$ref": "#/components/schemas/ItemPriceTypeEnum"
        itemId:
          type: string
          format: uuid
        itemType:
          "$ref": "#/components/schemas/ItemPriceItemTypeEnum"
        location:
          type: string
        locationUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - price
      - priceType
      - itemId
      - itemType
      - location
      - createdAt
      - updatedAt
      title: ItemPrice
...
    Manufacturer:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        code:
          type: string
        logo:
          type: string
        longName:
          type: string
        scRef:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
      additionalProperties: false
      required:
      - name
      - slug
      - createdAt
      - updatedAt
      - id
      title: Manufacturer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    Oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://fleetyards.net/oauth/authorize
          tokenUrl: https://fleetyards.net/oauth/token
          scopes:
            public: Public Access
            profile:read: Read Profile
            profile:write: Manage Profile
            hangar: Full Hangar Access
            hangar:read: Read Hangar
            hangar:write: Manage Hangar
            fleet: Full Fleet Access
            fleet:read: Read Fleet
            fleet:write: Manage Fleet
            openid: Authenticate your account
    OpenId:
      type: openIdConnect
      openIdConnectUrl: http://fleetyards.test/.well-known/openid-configuration
    SessionCookie:
      type: apiKey
      description: Session Cookie
      name: FLTYRD
      in: cookie